requirements-frozen.txt
acres==0.3.0
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asttokens==3.0.0
attrs==23.2.0
aws-cdk-lib==2.174.0
aws-cdk.asset-awscli-v1==2.2.223
aws-cdk.asset-kubectl-v20==2.1.4
aws-cdk.asset-node-proxy-agent-v6==2.1.0
aws-cdk.cloud-assembly-schema==39.2.20
aws-cdk.integ-tests-alpha==2.147.3a0
aws-solutions-constructs.aws-apigateway-lambda==2.62.0
aws-solutions-constructs.core==2.62.0
awswrangler==3.11.0
black==24.10.0
blinker==1.9.0
boto3==1.36.8
botocore==1.36.18
cachetools==5.5.2
cattrs==24.1.2
cdk-nag==2.28.159
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
cloudpickle==3.1.1
configobj==5.0.9
configparser==7.2.0
constructs==10.4.2
contourpy==1.3.2
cryptography==44.0.2
cycler==0.12.1
decorator==5.2.1
dill==0.4.0
docker==7.1.0
ecdsa==0.19.1
envs==1.4
etelemetry==0.3.1
Events==0.5
executing==2.2.0
extra-streamlit-components==0.1.80
Faker==26.0.0
faker_airtravel==0.4
fastapi==0.115.12
filelock==3.18.0
fitz==0.0.1.dev2
fonttools==4.57.0
git-remote-codecommit==1.17
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.5.2
ipython==8.29.0
isodate==0.6.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
jsii==1.106.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwcrypto==1.5.6
kiwisolver==1.4.8
looseversion==1.3.0
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mock==4.0.3
multiprocess==0.70.18
mypy-extensions==1.0.0
narwhals==1.35.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==1.26.4
omegaconf==2.3.0
opensearch-py==2.7.1
packaging==24.2
pandas==2.2.2
parso==0.8.4
pathlib==1.0.1
pathos==0.3.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
pox==0.3.6
ppft==1.7.7
prompt_toolkit==3.0.51
protobuf==5.29.4
prov==2.0.1
psutil==7.0.0
ptyprocess==0.7.0
publication==0.0.3
pure_eval==0.2.3
puremagic==1.28
pyarrow==18.1.0
pyasn1==0.4.8
pycognito==2024.5.1
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
pydeck==0.9.1
pydot==3.0.4
Pygments==2.19.1
PyJWT==2.10.1
PyMuPDF==1.25.5
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-jwt==4.1.0
pytz==2025.1
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==6.3.2
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rich==13.9.4
rpds-py==0.24.0
rsa==4.9.1
s3transfer==0.11.2
sagemaker==2.243.3
sagemaker-core==1.0.29
schema==0.7.7
scipy==1.15.2
setuptools==78.1.0
simplejson==3.20.1
six==1.16.0
smdebug-rulesconfig==1.0.1
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
starlette==0.46.2
streamlit==1.44.1
streamlit-cognito-auth==1.3.1
tblib==3.1.0
tenacity==9.1.2
termcolor==2.5.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
typeguard==2.13.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
uvicorn==0.34.2
watchdog==6.0.0
wcwidth==0.2.13
zipp==3.21.0
